Actulizado dias

+++ back/database/migrations/2025_08_23_202906_create_anulacions_table.php
@@ -13,26 +13,28 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('anulaciones', function (Blueprint $table) {
-            $table->id();
-            $table->date('fecha');
-            $table->string('cajero');
-            $table->decimal('monto', 10, 2);
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('user_autoriza_id')->nullable();
-            $table->unsignedBigInteger('user_anulacion_id')->nullable();
-            $table->string('motivo')->nullable();
-            $table->unsignedBigInteger('sale_id')->nullable();
-            $table->string('seccion')->nullable();
-            $table->string('detalle')->nullable();
-            $table->string('estado')->default('Pendiente');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('user_autoriza_id')->references('id')->on('users')->onDelete('set null');
-            $table->foreign('user_anulacion_id')->references('id')->on('users')->onDelete('set null');
-            $table->foreign('sale_id')->references('id')->on('sales')->onDelete('set null');
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('anulaciones')) {
+            Schema::create('anulaciones', function (Blueprint $table) {
+                $table->id();
+                $table->date('fecha');
+                $table->string('cajero');
+                $table->decimal('monto', 10, 2);
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('user_autoriza_id')->nullable();
+                $table->unsignedBigInteger('user_anulacion_id')->nullable();
+                $table->string('motivo')->nullable();
+                $table->unsignedBigInteger('sale_id')->nullable();
+                $table->string('seccion')->nullable();
+                $table->string('detalle')->nullable();
+                $table->string('estado')->default('Pendiente');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('user_autoriza_id')->references('id')->on('users')->onDelete('set null');
+                $table->foreign('user_anulacion_id')->references('id')->on('users')->onDelete('set null');
+                $table->foreign('sale_id')->references('id')->on('sales')->onDelete('set null');
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
